Adição de componentes e estilos para um design mais moderno

+++ frontend/src/components/DelButton.tsx
@@ -1,15 +1,18 @@
 import { deleteProduct } from "@/services/api";
 import { Button } from "./ui/button";
+import { Trash2 } from "lucide-react";
 
 const DelButton = ( { id } : { id: string }) => {
 
     const handleDelete = async () => {
-        await deleteProduct(id);
-        alert('Produto Removido!');
+        if (confirm('Voce realmente deseja deletar esse produto?')) {
+            await deleteProduct(id);
+            alert('Produto Removido!');
+        }
     }
 
     return (
-        <Button onClick={handleDelete} className="bg-red-500 hover:bg-red-700 hover:cursor-pointer text-white font-bold p-2">Deletar</Button>
+        <Button onClick={handleDelete} variant="destructive" className="font-medium flex-1 hover:cursor-pointer"> <Trash2 className="h-4 w-4" /> Deletar</Button>
     )
 }
